feature: 任务执行去除无用信息，优化头部导航栏
